Fix an issue disallowing a DailyVoiceClient to be recreated.
With this fix, you still can only have one DailyVoiceClient at a time,
with the most recent one working. However, instead of throwing an error
when trying to create a new one, it will disable the existing one.

This avoids issues in React when the DailyVoiceClient belongs to a
compoenent that comes and goes.